Making a simple impl of `in-radius` and using the old one selectively
Admittedly, this is for the sake of making the Tortoise equivalent
smaller.  The small, Tortoise-based version, however, outputs the agents
in a different order than this branch formerly was.  Thus, I'm rewriting
the implementation over here to be similarly concise and clear and
productive of the same ordering by default, and allowing for the
"optimized" version--which is, in many cases, not actually an
optimization--to be used when compiler optimizations are turned on and
the input agentset is `turtles`, `patches`, or a breed turtleset (which
it is optimized for, since `contains` on those is O(1)).