Feature: Serialize `jump` pen trails in a desirable way
Yes, I know about the similar code in `TrailDrawer`; I don't care about
it, since it's implemented in such a manic way.  I don't intend to
reimplement that approach in the Tortoise engine, nor to try and fix it.
I need this functionality in Tortoise, so I simply hooked up this engine
to produce the same results.

I'm aware that I just kind of hacked my solution into `Turtle`, but that
file is much too far gone for me to have a chance of saving it in any
reasonable amount of time.  So off with its head!